requirements/base.txt
Django>1.8
wsgiref==0.1.2
South>=0.8.1
Unipath>=1.0
psycopg2
django-extensions
flake8
django_compressor
dj-database-url
django-decouple
djangorestframework
django-filter
django-markdown
django-haystack
requests
django-taggit
whoosh
django-allauth
django-mptt
djrill
django-gravatar2
django-uuslug
django-embed-video
